Second iteration of changing docs
