+++ fastparse/shared/src/main/scala/fastparse/Api.scala
@@ -1,4 +1,7 @@
 package fastparse
+import java.nio.ByteBuffer
+import java.nio.ByteOrder._
+
 import language.experimental.macros
 import fastparse.parsers.{Intrinsics, Terminals}
 import fastparse.Utils.HexUtils
@@ -127,3 +130,29 @@ object allByte extends ByteApi{
   val AnyDword = P( AnyByte.rep(exactly=4) )
 }
 object emptyByteApi extends ByteApi
+
+object ByteUtils {
+
+  import allByte._
+
+  trait ByteFormat {
+    def wrapByteBuffer(byteSeq: ByteSeq): ByteBuffer
+
+    val AnyWordI = P(AnyWord.!).map(wrapByteBuffer(_).getShort & 0xffff)
+    val AnyDwordI = P(AnyDword.!).map(wrapByteBuffer(_).getInt)
+    // TODO Dword should be unsigned, but the only option is to change it to Long, what seems quite bad
+
+    def repeatWithSize[T](p: Parser[T], sizeParser: Parser[Int] = AnyWordI): Parser[Seq[T]] =
+      P( sizeParser.flatMap(l => p.rep(exactly = l)) )
+  }
+
+  object LE extends ByteFormat {
+    // Little Endian format
+    def wrapByteBuffer(byteSeq: ByteSeq): ByteBuffer = ByteBuffer.wrap(byteSeq).order(LITTLE_ENDIAN)
+  }
+
+  object BE extends ByteFormat {
+    // Big Endian format
+    def wrapByteBuffer(byteSeq: ByteSeq): ByteBuffer = ByteBuffer.wrap(byteSeq).order(BIG_ENDIAN)
+  }
+}
